Add new job script generation method, clean up code
